Add agency endpoints, and superadmin auth
