+++ controllers/authController.js
@@ -87,3 +87,50 @@ exports.refreshToken = (req, res) => {
     return res.status(403).json({ message: 'Invalid refresh token' });
   }
 };
+
+exports.superadminLogin = async (req, res) => {
+  try {
+    const { email, password } = req.body;
+
+    // Check against environment variables
+    const superadminEmail = process.env.SUPERADMIN_EMAIL;
+    const superadminPassword = process.env.SUPERADMIN_PASSWORD;
+
+    if (!superadminEmail || !superadminPassword) {
+      return res.status(500).json({ message: 'Superadmin credentials not configured' });
+    }
+
+    // Validate credentials
+    if (email !== superadminEmail || password !== superadminPassword) {
+      return res.status(400).json({ message: 'Invalid credentials' });
+    }
+
+    // Create JWT token for superadmin
+    const payload = { 
+      userId: 'superadmin', 
+      role: 'superadmin',
+      email: superadminEmail
+    };
+    
+    const accessToken = jwt.sign(payload, JWT_SECRET, { 
+      expiresIn: process.env.JWT_EXPIRATION || '15m' 
+    });
+
+    const refreshToken = jwt.sign(payload, process.env.JWT_REFRESH_SECRET, { 
+      expiresIn: process.env.JWT_REFRESH_EXPIRATION || '15d' 
+    });
+
+    res.json({ 
+      accessToken, 
+      refreshToken, 
+      user: { 
+        id: 'superadmin', 
+        email: superadminEmail, 
+        role: 'superadmin' 
+      } 
+    });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ message: 'Login failed' });
+  }
+};
